Move pathProvidesExec back to run runner
